second_commit

+++ LABASSIGN02/src/Q4.java
@@ -1,10 +1,11 @@
 public class Q4 {
+//Kareena 19105112
     public static void main(String[] args) {
-        double population = 80000;
+        int population = 80000;
         int years=0;
         while(population <=150000){
             years++;
-            population=population*1.05;// Adding 5% of current population to previous value
+            population=(int)(population*1.05);// Adding 5% of current population to previous value
            // System.out.println(population);
         }
         System.out.println("Time taken for Population to exceed 150000: " + years+ " years");
